Use distro-configured debuginfod server
Fedora and CentOS configure a debuginfod server by default, whereas RHEL
and Amazon Linux do not.  The new `dist_debuginfod_url` macro sets the
server URL in a central place.  Use that value, if defined, to configure
with a default server.  If unset, do not configure one.

This ensures we have a unified spec file for all distro variants, and
distro-specific config resides in the macros.dist file.

With this change, we do not need the recently-added bcond for
with_debuginfo_url, so drop it.

Signed-off-by: Amit Shah <amitshah@fedoraproject.org>